cleanup(publish)!: remove --contents flag - this was a legacy lerna feature which made the publishing flow much more complex than it has to be. we no longer use this, so remove it from publisher, simplifying the flow. - has a side effect of executing the prepublish scripts right before publishing. (instead of a special step at the beginning of publish)
