--- test/fixtures/dist-dir/.gitignore
@@ -1,1 +1,0 @@
-npm
